[CVW-026] Feat, 항상 성공, 실패하는 Stub객체 생성
